+++ build.gradle
@@ -1,9 +1,10 @@
-group 'pl.kamrar.zarcie'
+group 'pl.kamrar.gjmh'
 version '0.0.1-SNAPSHOT'
 
 apply plugin: 'java'
 apply plugin: 'idea'
 apply plugin: 'spring-boot'
+apply plugin: 'nebula.provided-base'
 
 sourceCompatibility = 1.8
 targetCompatibility = 1.8
@@ -14,6 +15,7 @@ buildscript {
     }
     dependencies {
         classpath("org.springframework.boot:spring-boot-gradle-plugin:1.3.3.RELEASE")
+        classpath("com.netflix.nebula:gradle-extra-configurations-plugin:3.0.3")
     }
 }
 
@@ -28,6 +30,7 @@ dependencies {
     compile('io.vertx:vertx-rx-java:3.2.1')
     testCompile('io.vertx:vertx-unit:3.2.1')
     testCompile('junit:junit:4.12')
+    provided('org.projectlombok:lombok:1.16.8')
 }
 
 jar {
