+++ settings.gradle
@@ -1,2 +1,2 @@
-rootProject.name = 'zarcie'
+rootProject.name = 'gjmh'
