+++ src/main/java/pl/kamrar/gjmh/verticle/ComplexStaticServer.java
@@ -1,9 +1,9 @@
-package pl.kamrar.zarcie.example.complex.verticle;
+package pl.kamrar.gjmh.verticle;
 
 import io.vertx.rxjava.ext.web.Router;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
-import pl.kamrar.zarcie.example.complex.verticle.helper.DefaultVerticle;
+import pl.kamrar.gjmh.verticle.helper.DefaultVerticle;
 
 /*
     Complex use of Vert.x server with multiple endpoints. Probably default our dev
